openAiwithPython/pyproject.toml
[project]
name = "openaiwithpython"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "colorama>=0.4.6",
    "openai==0.28",
    "prompt-toolkit>=3.0.0,<4.0.0",
    "python-dotenv>=1.0.0",
]